backend-python/requirements.txt
requests
msal
python-dotenv
botbuilder-core
botbuilder-schema
botframework-connector
aiohttp
